cargo clippy fixes

+++ src/image/lowlevel.rs
@@ -285,7 +285,7 @@ unsafe extern "C" fn finished_callback(converter: *mut wkhtmltoimage_converter, 
         // call and remove this converter's FINISHED_CALLBACK
         let mut callbacks = FINISHED_CALLBACKS.lock().unwrap();
         if let Some(mut cb) = callbacks.remove(&id) {
-            cb(val as i32);
+            cb(val);
         }
     }
 }
